Removed JCommander completely
The `Submit` class now uses the new pico-based `Main` program. So no more JCommander.
