+++ CONTRIBUTING.md
@@ -239,7 +239,7 @@ are all synonyms):
 
     ./gradlew shadowJar
 
-This will produce an assembly jar at `./flux-cli/build/libs/flux-cli-0.2.0-all.jar`.
+This will produce an assembly jar at `./flux-cli/build/libs/flux-cli-1.0-SNAPSHOT-all.jar`.
 
 You can now run any CLI command via spark-submit. This is an example of previewing an import of files - change the value
 of `--path`, as an absolute path is needed, and of course change the value of `--master` to match that of your Spark
@@ -247,7 +247,7 @@ cluster:
 
 ```
 $SPARK_HOME/bin/spark-submit --class com.marklogic.flux.cli.Submit \
---master spark://NYWHYC3G0W:7077 flux-cli/build/libs/flux-cli-0.2.0-all.jar \
+--master spark://NYWHYC3G0W:7077 flux-cli/build/libs/flux-cli-1.0-SNAPSHOT-all.jar \
 import-files --path /Users/rudin/workspace/flux/flux-cli/src/test/resources/mixed-files --preview 5 --preview-drop content
 ```
