+++ flux-cli/build.gradle
@@ -23,10 +23,8 @@ dependencies {
     // The rocksdbjni dependency weighs in at 50mb and so far does not appear necessary for our use of Spark.
     exclude module: "rocksdbjni"
   }
-  implementation 'info.picocli:picocli:4.7.6'
-  implementation "org.jcommander:jcommander:1.83"
-
   implementation "com.marklogic:marklogic-spark-connector:2.2-SNAPSHOT"
+  implementation "info.picocli:picocli:4.7.6"
 
   // hadoop-aws depends on aws-java-sdk-bundle, which clocks in as a single 380mb jar. We only need the S3 portion of
   // the AWS SDK, so the bundle is excluded in favor of the S3 API in the SDK.
@@ -55,8 +53,8 @@ dependencies {
   // For configuring two-way SSL in tests.
   testImplementation "com.marklogic:ml-app-deployer:4.7.0"
 
-  shadowDependencies "org.jcommander:jcommander:1.83"
   shadowDependencies "com.marklogic:marklogic-spark-connector:2.2-SNAPSHOT"
+  shadowDependencies "info.picocli:picocli:4.7.6"
 }
 
 javadoc {
@@ -82,7 +80,7 @@ distributions {
 }
 
 application {
-  mainClass = "com.marklogic.flux.cli.PicoMain"
+  mainClass = "com.marklogic.flux.cli.Main"
   // Removes warnings due to Spark performing "illegal reflective access".
   applicationDefaultJvmArgs = ['--add-opens', 'java.base/java.nio=ALL-UNNAMED', '--add-opens', 'java.base/java.net=ALL-UNNAMED',
                                '--add-opens', 'java.base/java.lang=ALL-UNNAMED', '--add-opens', 'java.base/java.util=ALL-UNNAMED',
